✨ devide file
